requirements.txt
flask
flask-cors
requests
oss2
pywebview
pyinstaller
httpx